move sort_accept_offers to pyramid.config.util
keeping this code isolated for now as it's kind of crazy
